calendar component modularized
